requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
codegen==1.0
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-appconf==1.0.4
django-bootstrap-form==3.4
django-countries==6.1.2
django-crispy-forms==1.9.0
django-datatables-view==1.19.1
django-debug-permissions==1.0.0
django-debug-toolbar==2.2
django-easy-pdf==0.1.1
django-filter==2.2.0
django-happenings==0.3.3
django-heroku==0.3.1
django-ipware==2.1.0
django-makemessages-xgettext==0.1.1
django-six==1.0.4
django-staticfiles==1.2.1
django-tinymce==2.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
mailer==0.8.1
Markdown==3.2.1
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.3
patterns==0.3
pefile==2019.4.18
Pillow==7.0.0
psycopg2==2.8.4
py==1.4.22
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyInstaller==3.6
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pytest==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.1
pywin32-ctypes==0.2.0
reportlab==3.5.42
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
spotipy==2.11.2
sqlparse==0.3.0
stripe==2.44.0
txaio==20.1.1
uno==0.3.3
urllib3==1.25.8
validators==0.14.3
virtualenv==20.0.20
webdriver-manager==2.3.0
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.4
XlsxWriter==1.2.8
zope.interface==5.0.0
djangorestframework-simplejwt==4.4.0
django-rest-auth==0.9.5